+++ components/collector/tests/source_collectors/azure_devops/test_pipeline_duration.py
@@ -7,7 +7,6 @@ class AzureDevopsPipelineDurationTest(AzureDevopsPipelinesTestCase):
     """Unit tests for the CI-pipeline duration metric."""
 
     METRIC_TYPE = "pipeline_duration"
-    METRIC_ADDITION = "max"
 
     def create_pipeline_run(
         self,