+++ components/collector/tests/source_collectors/gatling/test_source_up_to_dateness.py
@@ -13,7 +13,6 @@ class GatlingSourceUpToDatenessTest(GatlingTestCase):
     """Unit tests for the Gatling source up-to-dateness collector."""
 
     METRIC_TYPE = "source_up_to_dateness"
-    METRIC_ADDITION = "max"
 
     async def test_no_transactions(self):
         """Test that the age is 0 if there is no date in the HTML."""