+++ components/collector/tests/source_collectors/testng/test_source_up_to_dateness.py
@@ -13,7 +13,6 @@ class TestNGSourceUpToDatenessTest(TestNGCollectorTestCase):
     """Unit tests for the source up-to-dateness collector."""
 
     METRIC_TYPE = "source_up_to_dateness"
-    METRIC_ADDITION = "max"
 
     async def test_source_up_to_dateness(self):
         """Test that the source age in days is returned."""